+++ mem0/client/main.py
@@ -11,6 +11,7 @@ import requests
 from mem0.memory.setup import get_user_id, setup_config
 from mem0.memory.telemetry import capture_client_event
 from mem0.client.project import Project, AsyncProject
+from mem0.client.utils import api_error_handler
 
 logger = logging.getLogger(__name__)
 
@@ -18,29 +19,6 @@ warnings.filterwarnings("default", category=DeprecationWarning)
 
 # Setup user config
 setup_config()
-
-
-class APIError(Exception):
-    """Exception raised for errors in the API."""
-
-    pass
-
-
-def api_error_handler(func):
-    """Decorator to handle API errors consistently."""
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except httpx.HTTPStatusError as e:
-            logger.error(f"HTTP error occurred: {e}")
-            raise APIError(f"API request failed: {e.response.text}")
-        except httpx.RequestError as e:
-            logger.error(f"Request error occurred: {e}")
-            raise APIError(f"Request failed: {str(e)}")
-
-    return wrapper
 
 
 class MemoryClient:
@@ -631,7 +609,7 @@ class MemoryClient:
             APIError: If the API request fails.
             ValueError: If org_id or project_id are not set.
         """
-        logger.warning("This method is going to be deprecated soon. Please use the client.project.get() method instead.")
+        logger.warning("get_project() method is going to be deprecated soon. Please use the client.project.get() method instead.")
         if not (self.org_id and self.project_id):
             raise ValueError(
                 "org_id and project_id must be set to access instructions or "
@@ -676,7 +654,7 @@ class MemoryClient:
             APIError: If the API request fails.
             ValueError: If org_id or project_id are not set.
         """
-        logger.warning("This method is going to be deprecated soon. Please use the client.project.update() method instead.")
+        logger.warning("update_project() method is going to be deprecated soon. Please use the client.project.update() method instead.")
         if not (self.org_id and self.project_id):
             raise ValueError(
                 "org_id and project_id must be set to update instructions or "
@@ -1449,7 +1427,7 @@ class AsyncMemoryClient:
             APIError: If the API request fails.
             ValueError: If org_id or project_id are not set.
         """
-        logger.warning("This method is going to be deprecated soon. Please use the client.project.get() method instead.")
+        logger.warning("get_project() method is going to be deprecated soon. Please use the client.project.get() method instead.")
         if not (self.org_id and self.project_id):
             raise ValueError("org_id and project_id must be set to access instructions or categories")
 
@@ -1487,7 +1465,7 @@ class AsyncMemoryClient:
             APIError: If the API request fails.
             ValueError: If org_id or project_id are not set.
         """
-        logger.warning("This method is going to be deprecated soon. Please use the client.project.update() method instead.")
+        logger.warning("update_project() method is going to be deprecated soon. Please use the client.project.update() method instead.")
         if not (self.org_id and self.project_id):
             raise ValueError("org_id and project_id must be set to update instructions or categories")
