+++ mem0/client/project.py
@@ -6,6 +6,7 @@ import httpx
 from pydantic import BaseModel, Field
 
 from mem0.memory.telemetry import capture_client_event
+from mem0.client.utils import api_error_handler
 
 logger = logging.getLogger(__name__)
 
@@ -31,29 +32,6 @@ class ProjectConfig(BaseModel):
     class Config:
         validate_assignment = True
         extra = "forbid"
-
-
-class APIError(Exception):
-    """Exception raised for errors in the API."""
-    pass
-
-
-def api_error_handler(func):
-    """Decorator to handle API errors consistently."""
-    from functools import wraps
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except httpx.HTTPStatusError as e:
-            logger.error(f"HTTP error occurred: {e}")
-            raise APIError(f"API request failed: {e.response.text}")
-        except httpx.RequestError as e:
-            logger.error(f"Request error occurred: {e}")
-            raise APIError(f"Request failed: {str(e)}")
-
-    return wrapper
 
 
 class BaseProject(ABC):
